requirements.txt
beanie==1.26.0
guilded.py==1.13.2
pyyaml==6.0.1